Never let device replies count as client activity; clamp a stale scroll offset
Two device-reported bugs.

1. Every pane opened as "Mirroring another device" — the phone owned every
   session by default. PtyRoutes called noteClientInput() on EVERY inbound binary
   frame, and the arbiter is latest-active-wins, so any frame made that client the
   size governor. But an emulator answers the running program's queries by itself
   (cursor position, device attributes, colour/mode reports) over that same
   channel, so a phone that was merely mirroring seized the grid the moment a
   program probed the terminal. This is the server-side twin of the client-side
   hijack fixed in 03ac906: that stopped the phone forcing a resize, but the
   replies still reached the server and still counted as "the user is active
   here".

   The classifier moves to :clientServer (TerminalInputClassifier) so the clients
   and the server share one definition instead of restating it; PtyPresentation
   now delegates, keeping its API and tests. The replies are still written — the
   program is blocked on the answer — they just no longer record activity.

2. Crash while scrolling a mirror (a regression from the scroll fix in 230d37f):

     IllegalArgumentException: extRow=-3, mScreenRows=55, mActiveTranscriptRows=0
       at TerminalBuffer.externalToInternalRow <- TerminalRenderer.render <- onDraw

   mTopRow is clamped against the transcript length at the time of the gesture,
   but the transcript can shrink underneath it: a synthesized redraw starts with
   RIS + ED3 (clear scrollback), so a resync arriving while the user sits scrolled
   up leaves mTopRow pointing at rows that are gone. Before local scrolling was
   introduced the alt buffer never moved mTopRow, so this could not happen.
   Clamp at the read/render sites (clampTopRow, used by onDraw and getText) —
   which also covers the older ArrayIndexOutOfBoundsException from getText.

Verified: :server:compileKotlin, :web:compileKotlinJs, :androidApp:compileDebugKotlin,
PtyPresentationTest green.
